Changed quota from int to long to accomodate more reasonable quota sizes
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@21820 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/UserDatasetSession.java
@@ -124,7 +124,7 @@ public interface UserDatasetSession extends AutoCloseable {
    * @param userId
    * @return
    */
-  Integer getQuota(Long userId) throws WdkModelException;
+  Long getQuota(Long userId) throws WdkModelException;
 
   /**
    * Check if a user has a userId directory in the store.
